Moment navigation: factor translator out of closure
A context property provides the procedure used to export collected
moment-location data. This will allow for flexibility in implementing moment
navigation and possible upstreaming of the translator definition into LilyPond.
For example, the translator could be added to the output def of a user \score
block, enabling navigation data to be generated on compile instead of on eval,
with full flexibility of a custom defined \score. Or the callback could write
the nav data to a file or port instead of collecting it in a Scheme variable,
for use by programs other than Emacs.