Moment navigation: major rewrite of data generation
Instead of generating navigation data via evaluation in the REPL, rely on
compilation, since users will do this anyway and \score blocks automatically
handle the definition of which variables to include together.

When a \navigation output-def is included in a \score, its navigation data will
be exported to a file within a subdirectory of the LilyPond output directory
named .nav. Emacs will watch this folder and whenever a new or changed nav file
is detected, generate overlays and navigation table.

Use a score-id lookup to handle the case of an input file containing music for
multiple separate scores.