Major rewrite: \navigation output-def
Instead of generating navigation data via a function call that constructs score
music on the fly, rely on compilation, since users will do this anyway and
\score blocks already specify which variables to include together. This avoids
editing software needing to figure out which variables go together and how.

When a \navigation output-def is included in a \score, its navigation data will
be exported to a file within a subdirectory of the LilyPond output directory
named .nav. An editor should watch this folder and whenever a new or changed nav
file is detected, generate overlays and navigation table.

Use a score-id lookup to handle the case of an input file containing music for
multiple separate scores.