Use input file location for .nav folder
There isn't a reasonable way for an editor to figure out where LilyPond's output
directory is, since it may be set via arbitrary Scheme code. As nav data is
semantically associated with the input file, not the output, it seems reasonable
for .nav output to ignore user settings for LilyPond output location.