Simplified rewrite: single export table without splitting segments
The old approach of using a lookup table for navigation is brittle, because the
locations in the table become incorrect quickly during editing. It is awkward to
require recompiling in the middle of a vertical edit, since such edits may leave
the score in a rhythmically inconsistent state.

In Emacs, the natural solution is to use overlays for the lookup, since overlays
stay with the correct portion of text. As a consequence, there is no longer a
need to export separately formatted data for the lookup table and metadata.

This means also that the exported data no longer splits sequential segments, as
the import code already needs to iterate over all the data while populating
overlays.

How this would affect non-Emacs editors is a question, but at the moment it
seems prudent to prioritize the best version of the feature for Emacs, and let
other editors use that as a model.